backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
python-multipart>=0.0.9
openai>=1.0.0
